--- index.html
@@ -1,101 +1,0 @@
-<!-- 包括了首页导航栏、关于我们、团队成员、项目展示和联系我们等部分。
-导航栏内容需要修改 -->
-
-<html lang="zh-CN">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>团队介绍</title>
-    <link rel="stylesheet" href="styles.css" />
-  </head>
-  <body>
-    <header>
-      <nav>
-        <ul>
-          <li><a href="#about">关于我们</a></li>
-          <li><a href="#team">关于团队</a></li>
-          <li><a href="#projects">项目</a></li>
-          <li><a href="#partners">指导老师</a></li>
-          <li><a href="#contact">联系我们</a></li>
-        </ul>
-      </nav>
-    </header>
-
-    <section id="about">
-      <h2>关于我们</h2>
-      <p>我们是一个专注于技术创新的团队，致力于开发高质量的软件解决方案。</p>
-      <video width="320" height="240" controls>
-        <source src="about_video.mp4" type="video/mp4" />
-        您的浏览器不支持 video 标签。
-      </video>
-    </section>
-
-    <section id="team">
-      <h2>团队成员</h2>
-      <div class="team-card">
-        <img src="member1.jpg" alt="张伟" />
-        <h3>张伟</h3>
-        <p>首席技术官</p>
-        <a href="https://github.com/zhangwei">GitHub</a>
-      </div>
-      <div class="team-card">
-        <img src="member2.jpg" alt="李娜" />
-        <h3>李娜</h3>
-        <p>项目经理</p>
-        <a href="https://github.com/linana">GitHub</a>
-      </div>
-      <!-- 可以根据需要添加更多团队成员 -->
-    </section>
-
-    <section id="projects">
-      <h2>项目</h2>
-      <div class="project-card">
-        <h3>项目A</h3>
-        <p>这是一个旨在解决特定问题的项目。</p>
-        <img src="project_a.jpg" alt="项目A" />
-      </div>
-      <div class="project-card">
-        <h3>项目B</h3>
-        <p>这是一个旨在提供高效解决方案的项目。</p>
-        <img src="project_b.jpg" alt="项目B" />
-      </div>
-      <!-- 可根据需要添加更多项目 -->
-    </section>
-
-    <section id="teachers">
-      <h2>指导老师</h2>
-      <p>我们与以下老师建立了合作关系：</p>
-      <div class="teacher-card">
-        <img src="teacher1.jpg" alt="指导老师1" />
-        <h3>张老师</h3>
-        <p>资深前端工程师</p>
-        <a href="https://github.com/zhangwei">GitHub</a>
-      </div>
-      <div class="teacher-card">
-        <img src="teacher2.jpg" alt="指导老师2" />
-        <h3>李老师</h3>
-        <p>资深后端工程师</p>
-        <a href="https://github.com/linana">GitHub</a>
-      </div>
-      <!-- 可根据需要添加更多指导老师 -->
-    </section>
-
-    <section id="contact">
-      <h2>联系我们</h2>
-      <p>如果您有任何问题或合作意向，请随时联系我们。</p>
-      <form action="#" method="post">
-        <label for="name">姓名:</label>
-        <input type="text" id="name" name="name" required />
-        <label for="email">电子邮件:</label>
-        <input type="email" id="email" name="email" required />
-        <label for="message">信息:</label>
-        <textarea id="message" name="message" required></textarea>
-        <button type="submit">提交</button>
-      </form>
-    </section>
-
-    <footer>
-      <p>版权所有 &copy; 创数据</p>
-    </footer>
-  </body>
-</html>
