+++ uchain/src/components/Navbar..js
@@ -1,4 +1,5 @@
 import React from "react";
+import "./Navbar.css"; // 引入样式文件
 
 const Navbar = () => {
   return (
